+++ src/fides/api/util/aws_util.py
@@ -22,6 +22,10 @@ def get_aws_session(
     if storage_secrets is None:
         # set to an empty dict to allow for more dynamic code downstream
         storage_secrets = {}
+
+    stored_region_name = storage_secrets.get("region_name")  # type: ignore
+    stored_assume_role_arn: Optional[str] = storage_secrets.get("assume_role_arn")  # type: ignore
+
     if auth_method == AWSAuthMethod.SECRET_KEYS.value:
         if not storage_secrets:
             err_msg = "Storage secrets not found for S3 storage."
@@ -33,12 +37,10 @@ def get_aws_session(
             aws_secret_access_key=storage_secrets[
                 StorageSecrets.AWS_SECRET_ACCESS_KEY.value  # type: ignore
             ],
-            region_name=storage_secrets.get("region_name"),  # type: ignore
+            region_name=stored_region_name,
         )
     elif auth_method == AWSAuthMethod.AUTOMATIC.value:
-        session = Session(
-            region_name=storage_secrets.get("region_name"),  # type: ignore
-        )
+        session = Session(region_name=stored_region_name)
         logger.info("Successfully created automatic session")
     else:
         logger.error("AWS auth method not supported: {}", auth_method)
@@ -48,20 +50,13 @@ def get_aws_session(
     sts_client = session.client("sts")
     sts_client.get_caller_identity()
 
-    if assume_role_arn:
+    target_assume_role_arn = assume_role_arn or stored_assume_role_arn
+    if target_assume_role_arn:
         try:
-            response = sts_client.assume_role(
-                RoleArn=assume_role_arn, RoleSessionName="FidesAssumeRoleSession"
-            )
-            temp_credentials = response["Credentials"]
-            logger.info(
-                f"Assumed role {assume_role_arn} and got temporary credentials."
-            )
-            return Session(
-                aws_access_key_id=temp_credentials["AccessKeyId"],
-                aws_secret_access_key=temp_credentials["SecretAccessKey"],
-                aws_session_token=temp_credentials["SessionToken"],
-                region_name=storage_secrets.get("region_name"),  # type: ignore
+            return get_assumed_role_session(
+                target_assume_role_arn,
+                sts_client,
+                stored_region_name,
             )
         except ClientError as error:
             logger.exception(
@@ -70,6 +65,22 @@ def get_aws_session(
             raise
     else:
         return session
+
+
+def get_assumed_role_session(
+    assume_role_arn: str, sts_client: Any, region_name: Optional[str] = None
+) -> Session:
+    response = sts_client.assume_role(
+        RoleArn=assume_role_arn, RoleSessionName="FidesAssumeRoleSession"
+    )
+    temp_credentials = response["Credentials"]
+    logger.info(f"Assumed role {assume_role_arn} and got temporary credentials.")
+    return Session(
+        aws_access_key_id=temp_credentials["AccessKeyId"],
+        aws_secret_access_key=temp_credentials["SecretAccessKey"],
+        aws_session_token=temp_credentials["SessionToken"],
+        region_name=region_name,
+    )
 
 
 def get_s3_client(